Added draft of running rector in multiple workdirs
