+++ src/Domain/FileSystem/WorkingDirectory.php
@@ -4,9 +4,21 @@ declare(strict_types=1);
 
 namespace PHPMate\Domain\FileSystem;
 
-interface WorkingDirectory
+final class WorkingDirectory
 {
-    public function getAbsolutePath(): string;
+    public function __construct(
+        private string $directory
+    ) {}
 
-    public function fileExists(string $file): bool;
+
+    public function getAbsolutePath(): string
+    {
+        return $this->directory;
+    }
+
+
+    public function fileExists(string $file): bool
+    {
+        return file_exists($this->directory . '/' . $file);
+    }
 }
